requirements.txt
anyio==3.7.0
click==8.1.3
colorlog==6.7.0
exceptiongroup==1.1.1
fastapi==0.98.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
psycopg2==2.9.6
pydantic==1.10.9
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0

passlib~=1.7.4
PyJWT~=2.7.0